Move ETag to a separate module
